Prevent plugins from changing mouse cursor during middle-click autoscroll.
Cursor updates were suppressed in EventHandler, but this doesn't
work for plugins which call SetCursorForPlugin directly on ChromeClient.
This was a preexisting bug, but it got worse with the recent rewrite
because the bug used to be mostly covered up by the old implementation
pushing cursor updates at 60fps.

BUG=735865

Review-Url: https://codereview.chromium.org/2961003002
Cr-Original-Commit-Position: refs/heads/master@{#483259}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: be5aadf1dfe860773906916f583270793e45db3d