[LayoutNG] Move bfc_offset_, end_margin_sturt_ to NGLayoutResult
This is in preparation for relayout zero-block-size floats quirk.

For an aborted layout we don't want the NGLayoutResult to have a Fragment
associated with it, but we do want a BfcOffset, and EndMarginStrut.

This also cleans up double vector for out-of-flow positioned descendants, while
I was there.

BUG=635619

Review-Url: https://codereview.chromium.org/2960233002
Cr-Original-Commit-Position: refs/heads/master@{#483422}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a0782d6a6407c627004bac723cf73751d8fec130