leveldb: Add DBTracker for exposing databases to Chrome's memory-infra.
This CL adds DBTracker class that keeps track of all opened databases
and reports them to memory-infra (part of Chrome tracing). For database
to be tracked and reported it needs to be opened with leveldb_env::OpenDB()
function, which is a drop-in replacement for leveldb::DB::Open(). Next CL
will change all leveldb::DB::Open() invocations and add a presubmit check
to prevent adding new ones.

BUG=711518

Review-Url: https://codereview.chromium.org/2855953002
Cr-Original-Commit-Position: refs/heads/master@{#483507}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 542a7c89b3e856ef642d3ff6352e661717172395