Refactor NGInlineNode::CollectInlines (both overloads)
This patch does the following refactoring:

1. Remove the parameter NGInlineNode::CollectInlines(LayoutBlockFlow*) because
it is always equal to GetLayoutBlockFlow().

2. Change NGInlineNode::CollectInlines(LayoutBlockFlow*, NGInlineItemsBuilder*)
to a local function of the .cc file, because it does not access any private
member of NGInlineNode.

3. Move call sites of NGInlineItemsBuilder::EnterBlock and ExitBlock into the
above local function, so that it is easier to reuse this function to build
offset mapping.

Step 3 is a preparation for crrev.com/2963093002

BUG=699017
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:linux_layout_tests_layout_ng

Review-Url: https://codereview.chromium.org/2965573003
Cr-Original-Commit-Position: refs/heads/master@{#483585}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8cbc86dacc8e15bddeb0bae234ebb00fcc7550af