Implement DocumentTimeline constructor
This patch adds a constructor to DocumentTimeline to match the spec at
http://w3c.github.io/web-animations/#the-documenttimeline-interface. The
constructor enables the zero time of the timeline to be set at
something other than the navigationStart moment.

Changing the playback rate of the timeline (a functionality provided for
DevTools support) still modifies this zero time. The initial value for
the zero time is saved in the origin_time_ member.

BUG=674514,624639,600248

Review-Url: https://codereview.chromium.org/2960733004
Cr-Original-Commit-Position: refs/heads/master@{#483594}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7ee5b6192d46b7a25c3670c1aa8b10547e1057ef