Changes to XHR overrideMimeType function
This CL changes the functionality of the overrideMimeType() function
to bring it more in line with the XHR specs,
along with general handling of the MIME type and charset.
The final response charset is represented as a method instead of a variable,
to try and reflect the behavior described in the specs.

Before this CL, the browser took any input from overrideMimeType()
and set it as the Content-Type for the XHR response.
There was no process to ensure that the override MIME type was valid.
This is inconsistent with the XHR specs' expected behavior for
the overrideMimeType() method and usage of the override MIME type.
This change minimizes those inconsistencies by
 - validating MIME type input to overrideMimeType()
 - fixing the incorrect behavior of setting the response Content-Type
       as the override MIME type
This should bring Chrome more in line with the overrideMimeType specs.

https://xhr.spec.whatwg.org/#the-overridemimetype()-method
https://xhr.spec.whatwg.org/#response-body

Tests are updated; this change causes some tests to pass that failed before.
The content-type-header test fails now; it seems that failure is expected,
as far as I understand the specs.
I'll work on the overridemimetype-blob test more later.
For now, it's not a regression, just a slight change in output.

Bug: 699085
Change-Id: I9a5d6f6ca4475c23d0c13846f47ca88202f2c2f6
Reviewed-on: https://chromium-review.googlesource.com/547918
Commit-Queue: Austin James Ahlstrom <aahlstrom@google.com>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Reviewed-by: Takeshi Yoshino <tyoshino@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#483603}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 257d5b52af684edd0840a516519ba340e16ef24d