Assume WasCanceled() is false when |error_occurred| is false
When the Resource loading is canceled, it should cause ErrorOccurred()
to be true and thus the WasCanceled() check in ScriptLoader looks
no-op.
This CL replaces the WasCanceled() check with CHECK(), and a
following CL will remove the CHECK() if not failing.

Bug: 686281
Change-Id: I1e8eb3c74743ef8f69f267a901dd0f2ac8178e01
Reviewed-on: https://chromium-review.googlesource.com/556306
Commit-Queue: Hiroshige Hayashizaki <hiroshige@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#483755}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 95f17ec05931807d8bc7af8cdeab5792f8aadde1