Bounding boxes in the accessibility tree should always be relative to a box.
The accessibility tree stores bounding boxes relative to the nearest container,
which allows for efficient handling of transformations in the layout tree.
Previously we were allowing any box model object to be a container, but
this resulted in incorrect results for the static text child of an inline
element with position:relative. Fix this by requiring containers to be
boxes.

BUG=737130

Review-Url: https://codereview.chromium.org/2955303002
Cr-Original-Commit-Position: refs/heads/master@{#483823}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f3c32d8c2da3969d676bc6255224d779bda01ed2