[LayoutNG] Float behavior simplification before larger patch.
This is just an intermediate patch which simplifies some of the float
logic.

As we won't be "re-parenting" floats away from their actual parents
anymore a lot of the complex logic around the offsets becomes simpler.

Specifically we won't need the "paint-offset" calculation anymore.

This is just an initial patch to reduce the behaviour, it has some
regressions, but these should be fixed with the new code.

BUG=635619

Review-Url: https://codereview.chromium.org/2945213004
Cr-Original-Commit-Position: refs/heads/master@{#483826}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e65e2d1acb3c3bb037f0183bdd3db4cf35da45c2