Remove the current non-generic baseline before rebaselining it
This avoids unnecessary non-generic baselines which are the same as the
fallback baseline.

This is similar to "webkit-patch optimize-baselines" but avoids extra
baselines at the first place.

For example, this is useful for optimizing flag-specific baselines in
the following process:
1. We created flag-specific baseline for a test when it produced
   result different from the existing expectations.
2. Then some code change makes the test produce flag-specific result
   the same as the original existing expectations before step 1.
3. We use run-webkit-tests --new-flag-specific-baseline <test>.

Previously at step 3, we would overwrite the flag-specific baseline with
the new actual result. Now we just remove the flag-specific baseline.

BUG=660231

Change-Id: I713c9d729f60f88ff29e46f3b995433dcdf36ba1
Reviewed-on: https://chromium-review.googlesource.com/558380
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#484040}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7dbd719a57553e94a2d1dfc9414a8d013749fc54