Add method pr_for_chromium_commit to WPTGitHub, and use it in importer
Reason:
While the is_exportable function was using both Change Id and commit
position to find the PR, test_importer was using commit position
to find the PR. Because of https://crbug.com/737178, that sometimes
returns the wrong PR number, which caused some confusion.

This CL extracts the logic used to get the PR for a commit to avoid
repetition, and makes it so it's used in both is_exportable and the
importer.

Bug: 727923
Change-Id: I9e5dfbdc2f8805693e1aa01760cb64951a6dd908
Reviewed-on: https://chromium-review.googlesource.com/558200
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Jeff Carpenter <jeffcarp@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#484402}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 995f3e6f520b491f9489a8842e1b3beb592a993d