Limit the repeatability of repeatable list animations
The combination of CSS transition retargeting and properties that
animate as repeatable lists [0] can lead to scenarios where the number
of items in an animated value explodes to hundreds of thousands.

This is almost always unintentional and can be a source of system
lock ups for the user. This patch limits how much animations will
repeat two lists to be no more than 1000.

[0] https://www.w3.org/TR/css3-transitions/#animtype-repeatable-list

Bug: 739197
Change-Id: Ib919b6af821f5d8e7cb98789c38ca568a70de52a
Reviewed-on: https://chromium-review.googlesource.com/558899
Commit-Queue: Alan Cutter <alancutter@chromium.org>
Reviewed-by: Douglas Stockwell <dstockwell@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#484444}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8f9409bacee5b415ae2bc07c1a5d41bf9d5c23d4