Recalculate effective root scroller on removal.
When the Element currently set as the document.rootScroller is removed
from the Document, it shouldn't be able to be the effective root
scroller anymore (see README.md in core/page/scrolling for definitions).

Currently, the effective root scroller is generally recalculated after a
layout. This means removing an Element doesn't necessarily cause a
recalc. This patch explicitly induces a recalc when the rootScroller is
removed from the Document.

Bug: 505516
Change-Id: I3f151f77d62ec3d47508c00b396c7402ed49ba61
Reviewed-on: https://chromium-review.googlesource.com/559933
Reviewed-by: Steve Kobes <skobes@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#484637}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ddc753934758db1cccdfc5491af7b58735a5c6f2