media: Add HDCP policy check on MediaKeys (blink)
This API provides an API for JS application to check the key status
associated with a policy without the need to fetch an actual license.
This is critical for starting playback at higher resolution and hence a
better user experience.

This CL updates the Blink side to support the new API:
- Add IDL for MediaKeysPolicy.
- Add IDL for GetStatusForPolicy (on partial MediaKeys interface).
- Implement GetStatusForPolicy on the blink side.
- Add a run time flag EncryptedMediaHdcpPolicyCheck to control this API.

The Chromium side change will be implemented in the next CL.

BUG=709348

Change-Id: I61009d516aae2d29a3dc3536214560d0495d250c
Reviewed-on: https://chromium-review.googlesource.com/540784
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Reviewed-by: John Rummell <jrummell@chromium.org>
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#484848}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6de258a1803009bb76b7d1dbce0f99c45dd67d39