Use WebPrivatePtr in WebHTTPBody and WebSecurityOrigin.
The previous idiom was to define a subclass to wrap the type, and store
a raw pointer to the subclass. Rather than doing that, just use
WebPrivatePtr.

Bonus fix: WebHTTPBody's assignment operator was accidentally changed to
a no-op. Luckily it appears that this is never called...

Change-Id: I6301f410513b4cd93ca1fcb537d46dd87359bb7a
Reviewed-on: https://chromium-review.googlesource.com/565077
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#485510}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 449b5c274f3310b9a78440dd77feb317544e2193