Group all responsive animation tests together
Move tests of animations being responsive to style changes under the
existing animations/responsive test directory. The tests that are
already there are testing responsiveness in interpolations, so they are
moved to an interpolation/ subdirectory.

The tests in the svg-attribute-responsive directory are also moved
under responsive/interpolation/ in order to avoid creating a copy of
the responsive-test.js with different relative path imports.

BUG=623434

Review-Url: https://codereview.chromium.org/2973013002
Cr-Original-Commit-Position: refs/heads/master@{#485512}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0a1a9e4d3577a94effac93c81110444f1fb9b7a4