[LayoutNG] Abort a layout once the BFC offset is resolved.
This patch allows layouts to "abort" if they have resolved their BFC offset.

This occurs when there is a previous sibling which has an unpositioned float
inside of it.
We resolve the BFC offset, then abort the layout up to the point where we can
begin the relayout (a parent which knows its BFC offset).

The algorithm in this state works exactly as before, except that floats are positioned
at their calculated BFC offset.

BUG=635619

Review-Url: https://codereview.chromium.org/2954953002
Cr-Original-Commit-Position: refs/heads/master@{#486065}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f905d51bad504b4b43149637760b649aea7ddc24