Move raster work of createImageBitmap(SVG) to background thread
Right now in html canvas 2d's drawImage API, the image decoding are done
on the main thread, and this task could be heavy when the image is big.
What we hope is to give people an option to call createImageBitmap(SVG)
which does the image decoding in a background thread, and then canvas
drawImage API can draw the image bitmap which is already in a decoded
state.

Bug: 580202
Change-Id: I79cfd46b4e2f154e9fde111b932125a9745571a2
Reviewed-on: https://chromium-review.googlesource.com/558398
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Justin Novosad <junov@chromium.org>
Commit-Queue: Xida Chen <xidachen@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#486144}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7a1fd06507b3a2e4b33b521398e08147b4185a7d