[LayoutNG] Refactor of HandleInflow children.
This moves a bunch of code around and some additional comments.

Before the code for handling an inflow element was a little difficult to follow.
Now most of the code and complexities are inside HandleInflow. Functions which
this calls are now relatively simple.

BUG=635619

Review-Url: https://codereview.chromium.org/2975973002
Cr-Original-Commit-Position: refs/heads/master@{#486260}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 987ca52741e0e4f0399de57732e4afd2382871cc