Media Controls: record display and interaction will buttons and sliders for CTR.
Adding the necessary scaffolding to compute the CTR of media controls
elements with as much precision as possible. This is not computing CTR
using visibility to the user but hypothetical visibility on the page in
the sense that the element was displayed in the controls but maybe never
scrolled into view.

In order for an element to be marked as displayed, the element has to
have metadata downloaded or be set to not preload by the page or the
browser.

This CL also removes the CTR code for the download button and the cast
overlay button given that this code is more precise and more generic.

Bug: 651644
Change-Id: I86c2bc4d1a19ab99a211082d8b3546b79ea71144
Reviewed-on: https://chromium-review.googlesource.com/567194
Commit-Queue: Mounir Lamouri <mlamouri@chromium.org>
Reviewed-by: Jennifer Apacible <apacible@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#486741}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: aeb364f107e1c979ede259f90d8f83fe89be9969