Lazy initialize media controls when not using the Cast overlay.
The work to abstract the media controls from the HTMLMediaElement allows
us to do this without risks. However, when the Cast overlay has to be
used, the controls need to be created even if they are not used. The
code implementing this logic is part of the native controls so we can't
lazy initialise the controls the same way.

Bug: 731143
Change-Id: I2eb912ab6124c24000860007c1d6e6aa3300c440
Reviewed-on: https://chromium-review.googlesource.com/528176
Commit-Queue: Mounir Lamouri <mlamouri@chromium.org>
Reviewed-by: Anton Vayvod <avayvod@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#486992}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fcc3b11bbc8233e2f338a89ef7245c372dc58115