Edits to support the overrideMimeType-blob test
These changes are to bring Chrome's XHR implementation
in greater agreement with the XHR specs (see links).
This change makes it so that setting the content type
as defined in the XHR response behavior for blobs
is no longer dependent on a non-blank file path,
which is more consistent with the specs.

https://xhr.spec.whatwg.org/#the-overridemimetype()-method
https://xhr.spec.whatwg.org/#response-body

A change to the overridemimetype-blob test is added
based on communication with a spec editor (see link).
This is due to an inconsistency between two specs:
the MIME-sniffing standard and RFC-2045 (see links).

http://logs.glob.uno/?c=freenode%23whatwg&s=6+Jul+2017&e=6+Jul+2017
(see comments by yhirano, annevk, and GPHemsley)
https://mimesniff.spec.whatwg.org/#parsing-a-mime-type
https://tools.ietf.org/html/rfc2045#page-12

This change fixed four test failures on the WPT test,
but created one test failure (xhr-response-blob.html).
I changed the expectations file related to this test
to accommodate the change.

There are two 'FIXME' comments relevant to this test:
XMLHttpRequest.cpp, lines 1428-1429 and lines 1593-1595.
I don't think that they're necessary, and would consider
removing them, but I don't know best practice for that.

Bug: https://bugs.chromium.org/p/chromium/issues/detail?id=699085
Change-Id: I77919178b725792e914d6984cc3f5ebde654e5af
Reviewed-on: https://chromium-review.googlesource.com/563139
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Reviewed-by: Takeshi Yoshino <tyoshino@chromium.org>
Commit-Queue: Austin James Ahlstrom <aahlstrom@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#487412}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4a20bad558e648c8e7208443d0d4a80f68587ba2