Implement AccessibleNode.hasPopUp for Accessibility Object Model phase 1
This is the last AOM phase 1 property! I missed it earlier because it used
to be a boolean attribute but in ARIA 1.1 it was changed to a string.
This change just hooks up the AOM portion, but I filed bug 740137 to
track plumbing through the string value of haspopup rahter than treating
it as a boolean elsewhere.

BUG=680345,740137

Review-Url: https://codereview.chromium.org/2973973002
Cr-Original-Commit-Position: refs/heads/master@{#487773}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fee38fe5d637a5f0ea3d06708ec12a9981fce6bb