Plumb VSync / BeginFrameSource paused state through RenderWidgetHostViewAndroid.
OnAndroid, VSync processing may be paused for the browser compositor,
which has the effect of not sending down OnBeginFrame events to the
render process, essentially pausing both browser and renderer
compositing.

Sometimes the renderer compositor can get into a bad state where the
main thread is blocked waiting on a completion
(ProxyMain::BeginMainFrame::commit), but we don't meet the conditions
for ACTION_ACTIVATE_SYNC_TREE, so the completion is never triggered.

The fix is to plumb paused state from the browser to the renderer
process, so ACTION_ACTIVATE_SYNC_TREE can be force activated (when
paused), and we can unblock the main thread.

BUG=738443

Change-Id: I401d67bdd0f1054d5594575c8800d01a53e67f2b
Reviewed-on: https://chromium-review.googlesource.com/567029
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: Bo Liu <boliu@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Reviewed-by: Fady Samuel <fsamuel@chromium.org>
Commit-Queue: Bill Orr <billorr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#488059}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 372a12674b9744c5a7742c99009358601125d0fc