Change behavior of SpellChecker::GetSpellCheckMarkerTouchingSelection()
When I added this method, I had added some extra logic for the case where we
have a collapsed selection to make it find spell check markers the caret is
touching an endpoint of. This was to support the spell check menu on Android,
where tapping the endpoint of a word brings up the menu (on other platforms you
need to right-click in the interior of the word for spell check options to
appear in the context menu).

Upon further reflection, we decided we should put this logic somewhere else.
We're going to keep this method around though because we're going to make use of
it in ContextMenuClient::SelectMisspellingAsync().

I'm renaming this method to GetSpellCheckMarkerUnderSelection() to try to make
the name less misleading.

BUG=736181

Review-Url: https://codereview.chromium.org/2954763002
Cr-Original-Commit-Position: refs/heads/master@{#488476}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 37f9a3079589e85e24eaa5849235b7eedc9fd0ab