DevTools: add IO.resolveBlob
This adds a protocol method to obtain the UUID of a Blob based on a
remote object id.
The intended usage is in conjunction with IO.read that would be able
to read the contents of a Blob on the browser side.

Bug: 747508

Change-Id: I18b7395833265665d80493f611888285f9730458
Reviewed-on: https://chromium-review.googlesource.com/580353
Commit-Queue: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#488751}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: da6245e7c4b489a4dc73be76fcb54483add1eb2b