Increase sample rate of tests to 8192
These tests use a sample rate of 4096 which is supported by Chrome.
But the spec says the minimum required rate is 8000.  To be able to
upstream these tests, we need to increase the sample rate.  Make it
8192.

Bug: 704955
Test: AudioBufferSource/audiobuffersource-loop-grain-no-duration.html, AudioBufferSource/audiobuffersource-playbackrate-zero.html, AudioBufferSource/audiobuffersource-premature-loop-stop.html, AudioParam/audioparam-linearRamp-value-attribute.html
Change-Id: Ic63d0ed5df65cfecfc27e87df05460aa749d7600
Reviewed-on: https://chromium-review.googlesource.com/578228
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Commit-Queue: Raymond Toy <rtoy@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#489060}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 66638bfc76ac48ae57b53fc131cec926e5e8c2df