Read CSP/ReferrerPolicy on WorkerThead when script streaming is enabled
This is 9th patch of script streaming project split off from
https://chromium-review.googlesource.com/c/538477.
Design doc:
https://docs.google.com/document/d/1BbETxB2K1GeGUv4XIvGGforAJRRgYSNX5x4vPTGsCPU/edit#heading=h.ubqh39vcdj3n

This patch sets ContentSecurityPolicy and ReferrerPolicy to the document and
SWGlobalScope correctly when script streaming is enabled. When script streaming
is enabled, the main script is provided on the worker thread, so this patch
posts the information to the main thread after calling
InstalledScriptsManager::GetScriptData on the worker thread.

Bug: 683037
Change-Id: I2232804ae762e7886b171638443a1b08592e870c
Reviewed-on: https://chromium-review.googlesource.com/578972
Commit-Queue: Makoto Shimazu <shimazu@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#489603}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4cb4c97b708d54bea4b045c2519a2eb7017f900c