Release mouse pointer capture when it starts drag
This CL releases mouse pointer capture if there
was one in place when drag operation starts.
Note that this doesn't do anything for touch
drag operation as that needs to be solved
when we have a better coordination of gesture
and touch events.

BUG=706106, 720201

Review-Url: https://codereview.chromium.org/2973963003
Cr-Original-Commit-Position: refs/heads/master@{#490010}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 656f9b939150e9a40553757cea46d385bcb423d9