Fix exposure of interfaces with [SecureContext]
IDL interfaces with [SecureContext] applied are still visible in
insecure contexts. This was found when fixing [OriginTrialEnabled] and
[SecureContext] to work together correctly (crbug.com/695123). Specifically,
with [SecureContext] in the IDL, the interface is still exposed as a
constructor attribute on Window.

When encountered on previous CLs, the solution was to run layout tests
in secure contexts:
https://codereview.chromium.org/2832813002
https://chromium-review.googlesource.com/c/506292/

This CL fixes the bindings logic so that [SecureContext] attribute is
flagged on the generated constructor attributes. The result is that the
existing bindings templates will generate the appropriate checks for
secure context.

Bug: 695123, 634270
Change-Id: If6d37d4441ba06d4f2e04f4360ed28990a0e2547
Reviewed-on: https://chromium-review.googlesource.com/508328
Commit-Queue: Jason Chase <chasej@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#490132}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8e7aa98860b23dc0d395b8b8306b7cc382b2c0ba