Worker: Clear WorkerOrWorkletGlobalScope in PrepareForShutdownOnWorkerThread()
This is a cleanup CL and shouldn't change behavior.

Before this CL, WorkerThread clears WorkerOrWorkletGlobalScope in
PerformShutdownOnWorkerThread() to be called in the last of WorkerThread
termination. However, we don't have to keep it until the last function and it'd
be more natural to clear it up in PrepareForShutdownOnWorkerThread() that calls
WorkerOrWorkletGlobalScope::Dispose() etc.

This CL moves the clearance to PrepareForShutdownOnWorkerThread() and removes
outdated comments added before the Oilpan is enabled.

Bug: 710364
Change-Id: Ibdc36555129d2d67c9f9cab9f47fb8dd4d31334f
Reviewed-on: https://chromium-review.googlesource.com/590336
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#490292}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 125504917421425b5411cdd7198a4ad7919ee5f6