Show touch handle for newlines
We were incorrectly checking if a textbox was empty in
SelectionController. This led to touch handles not being shown when
tapping into a contenteditable div that had a newline. This CL changes
the way we check for empty textboxes.

Bug: 746917
Change-Id: I9b0494a1d8ea3df27de29746490a184f54b7a56f
Reviewed-on: https://chromium-review.googlesource.com/584167
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: Pedro Amaral <amaralp@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#490459}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5f6256d36f375c76fbb1a504fb5581ee202ccea0