Long tasks: replace settimeout with step_timeout in test and fix the test
Replace settimeout with set_timeout in the test longtask-in-parentiframe.html,
according to testharness-api.
The original test fails because the long task is fired before the observer is
created, so this CL fixes it by delaying the long task until the iframe is
loaded.

Bug: 745031
Change-Id: I89e0f0fab5ef0458c773f90ded428fc1a483913c
Reviewed-on: https://chromium-review.googlesource.com/579593
Reviewed-by: Shubhie Panicker <panicker@chromium.org>
Commit-Queue: Liquan Gu <maxlg@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#490831}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7ae79f6845ea45bb6bc691dec8614ce85b302e1f