Remove self_ref in Canvas2DLayerBridge
In the current implementation of MailboxInfo, we keep a RefPtr to
Canvas2DLayerBridge, and when the MailboxReleased function gets called,
we do a self_ref to keep the Canvas2DLayerBridge alive.

Now that the MailboxReleased function is called through a WTF::Bind to
a WeakPtr of Canvas2DLayerBridge, we don't need the self_ref anymore.
Also, we can remove the mailboxes_ that is kept in this class, we just
need to remove the associated resources.

Bug: None
Change-Id: I7f087a23a7fca32272d0ad05efb66f92d9eabba8
Reviewed-on: https://chromium-review.googlesource.com/579690
Commit-Queue: Xida Chen <xidachen@chromium.org>
Reviewed-by: Justin Novosad <junov@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#490960}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 645461c33c2eed8afd68176138898d0b7080d793