Remove FetchContext::CanRequest() call from ResourceLoader::DidReceiveResponse()
It's been wrong that this method was enclosed by the if with
cors_handling_by_resource_fetcher in its condition. CanRequest() is not
about CORS but CSP, etc.

However, this was not harmful. Since
ResourceLoader::WillFollowRedirect() just calls it on each redirect,
calling this in DidReceiveResponse() was just redundant.

So, remove it.

The only concern was that the CanRequest() uses the SecurityOrigin
on the ResourceLoaderOptions updated by HandleRedirect() on the last
redirect if any, and we remove that effect. But it looks actually
unnecessary.

Bug: 736308
Change-Id: I8d7b30a1eac6d7f5f6151a0385e7019ab815c86c
Reviewed-on: https://chromium-review.googlesource.com/581955
Commit-Queue: Takeshi Yoshino <tyoshino@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#491350}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e4f503e9b261ae419d4e4cba3ef38aaed492f046