css: Count usages of StyleSheetList's named getter that do not return null.
The anonymous named getter is not part of the spec, and we are trying to
figure out how much it is actually used. The existing counter is triggered
whenever code tries to access a named property, this second counter is
triggered only when a property is found and should give us a better idea of
when named properties are not accessed by accident.

Bug: 689687
Change-Id: I7f255593d2a998a83a62ca60d814751b938db0af
Reviewed-on: https://chromium-review.googlesource.com/596975
Reviewed-by: Eric Willigers <ericwilligers@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Philip Jägenstedt <foolip@chromium.org>
Commit-Queue: Raphael Kubo da Costa (rakuco) <raphael.kubo.da.costa@intel.com>
Cr-Original-Commit-Position: refs/heads/master@{#491378}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f7a75591a52c88a9bcd24b1d97d14606d9236977