Dispatch interfacerequest event in a new microtask
This change modifies the MojoInterfaceInterceptor so that the
'interfacerequest' event, fired when a request for the intercepted
interface is made, is dispatched in a new microtask instead of
synchronously.

This has the benefit of not reentering JavaScript to execute event
handlers synchronously which is both different from the usual case for
binding interfaces that are provided by other processes and avoids the
possibilty of attempting to dispatch this event to JavaScript from a
scope where script is forbidden.

Change-Id: I08b8c25a04da047381a3786af08b62979ca7b837
Reviewed-on: https://chromium-review.googlesource.com/594716
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Yuzhu Shen <yzshen@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#491552}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fe5c2f3d9c18a5c652b064d230ef3e8f1f9f903d