EQT: Skip steps when renderer is backgrounded
In this CL, the QueueingTimeEstimator only uses visible steps for
calculating Expected Queueing Time. A step is considered visible if
the renderer was never backgrounded throughout the step (excluding
possible changes at exact start or end of step). A window may thus be
made up of steps that are not contiguous in time. The objective of this
CL is to start reducing the massive concentration of the UMA stats for
Expected Queueing Time on the 0-1 ms bucket, because that is not very
useful and not a good indicator of latency.

Bug: chromium:729033
Change-Id: I3cfd4d3268d0f2858fc77de25c08d35a7a1cfc36
Reviewed-on: https://chromium-review.googlesource.com/587588
Commit-Queue: Nicolás Peña <npm@chromium.org>
Reviewed-by: Timothy Dresser <tdresser@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#491773}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6be7ecdcae29068ea6ac0f60e9ab880d67c13156