Allow interactive rebaselining in test-expectations.html
This patch adds the interactive rebaseline functionality to test-expectations.html,
so that after checking the test results, a selective rebaselining can be easier:

1. Click the "Select" button next to "Select tests to rebaseline"
2. A checkbox appears by each shown test
3. Check the checkboxes which needs rebaseline
4. Click the "Done" button
5. A dialog with some rebaseline script appears, with the script already copied
   to clipboard. Run the script [1] locally in your terminal in LayoutTests/ (or
   platform/xxx/, flag-specific/xxx/ if the tests are platform/flag dependent)
   to download the new expectation files to create a rebaseline patch.

The initial purpose of this patch is to help the the development of
TextIteratorNG [2], which creates a lot of unexpected failures. Some of them are
due to resolution of existing bugs (and hence, these tests should be rebaselined),
but others are due to new bugs (so these tests shouldn't be rebaselined). This
patch makes the selective rebaseline much easier.

[1] The script is currently Linux-only
[2] https://chromium-review.googlesource.com/c/580371

Bug: 
Change-Id: If869509f27f8d32e938f9778f636513045845a01
Reviewed-on: https://chromium-review.googlesource.com/597388
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Reviewed-by: Emil A Eklund <eae@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#491808}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: aca9436222ecf4124f47a7dec635762011007b12