Make a copy of String function in ReadableStream
The ReadableStream constructor uses the String function. However, the
implementation was failing to take a copy of the function. This meant
that the version from the Javascript environment was used. If it had
been modified then it could result in incorrect behaviour.

Use a copy of the original value of the String function in
ReadableStream.

This change also includes a regression test.

BUG=752177

Change-Id: I52f36244a9d87131219958837e3cee00a8fc5fb4
Reviewed-on: https://chromium-review.googlesource.com/601668
Reviewed-by: Takeshi Yoshino <tyoshino@chromium.org>
Commit-Queue: Adam Rice <ricea@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#491973}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: be5214ee56f8f91275210e297b961a5b836d94a3