[wrapper-tracing] Switch to conservative Dijkstra barrier
On assignment 'x.a = y' during incremental marking the Dijkstra barrier
suggests checking the color of 'x' and only mark 'y' if 'x' is marked.

Since checking 'x' is expensive in the current setting, as it requires
either a backpointer or expensive lookup logic due to large objects and
multiple inheritance, just assume that 'x' is black. We assume here that
since an object 'x' is referenced for a write, it will generally also be
alive in the current GC cycle.

As a side-effect this CL also
- Remove the need for explicit construction of TraceWrapperMember.
- Fixes swap() performance when tracing is off.
- Enable the use of more TraceWrapperMember where we previously used
  a manual write barrier.

Bug: 
Change-Id: Ic283a2eaac1323aa3019e08ba9fa330900bdd0aa
Reviewed-on: https://chromium-review.googlesource.com/598907
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Ulan Degenbaev <ulan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#492014}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cd39e5b0ea3c0da423e54066bd97571f82dffaad