Add HitTestRegionList to SubmitCompositorFrame
We need a way to send HitTestRegionList data to the HitTestAggregator
with each call to SubmitCompositorFrame.

This change adds HitTestRegionList as an optional argument to
SubmitCompositorFrame.  The HitTestRegionList data is currently
optional and empty - this will be added in future CLs.

Bug: 732399
Change-Id: I291bf67b36a3e3addf59a28fb64b58f2dc5527ce
Reviewed-on: https://chromium-review.googlesource.com/585294
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Justin Novosad <junov@chromium.org>
Commit-Queue: Gary Klassen <gklassen@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#492242}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d9e90afec475d912cca91cd0b42e7b3a564fff59