Avoid calling EnsureUAStyleForFullscreen() when exiting fullscreen
Because the document may be inactive, and when it isn't the UA style
sheet is not needed any longer, and should already be loaded.

The added test triggered the same DCHECK as discussed in the bug.

Bug: 747827
Change-Id: I36e862ff3a8086315d8be47169b5d5a2a9033a0f
Reviewed-on: https://chromium-review.googlesource.com/602272
Reviewed-by: Rune Lillesveen <rune@opera.com>
Commit-Queue: Philip Jägenstedt <foolip@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#492328}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a67e690ff85360414ce2db240ee5969b19595f67