Use uninitialized AudioBuffer for startRendering()
startRendering() will fill up the resulting AudioBuffer with the
rendered data so it's a waste to zero-initialize the buffer when it's
going to be filled up anyway.  Therefore, allow creation of an
AudioBuffer that is not zero-initialized.

Also, when creating and AudioBuffer from an AudioBus, there's no need
to initialize the AudioBuffer either because we're going to immediately
copy the bus contents to the buffer.

Bug: 736561
Test: none
Change-Id: Ibfe1c0d7da0fba62901445a5530026e6ff297b28
Reviewed-on: https://chromium-review.googlesource.com/586858
Commit-Queue: Raymond Toy <rtoy@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#492361}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 28b810692b54bb562b51b36de2c4701a49954498