ServiceWorker: Notify errors on sending installed scripts to renderer
This is 10th patch of script streaming project split off from
https://chromium-review.googlesource.com/c/538477.
Design doc:
https://docs.google.com/document/d/1BbETxB2K1GeGUv4XIvGGforAJRRgYSNX5x4vPTGsCPU/edit

This patch fixes error handling on ServiceWorkerInstalledScriptsSender. The
renderer can know the browser-side error by
- Disconnection of mojom::ServiceWorkerInstalledScriptsManager before
  TransferInstalledScript.
- Disconnection of each data pipe after TransferInstalledScript.

For telling the data transfer is completed successfully, this patch adds
body_size and meta_data_size to mojom::ServiceWorkerScriptInfo. If total amount
of received data is different from the size when the data pipe is closed, it
means an error happened during sending the body or meta data. In that case,
ThreadSafeScriptContainer returns nullptr. That results in fallback to the
ResourceLoader path (a follow-up patch will change this to result in error
instead). If an error happens for the main script, the worker will be
terminated.

Bug: 683037
Change-Id: I3fddcab920d272be0187d3b48380aa9136ab0f4c
Reviewed-on: https://chromium-review.googlesource.com/579307
Commit-Queue: Makoto Shimazu <shimazu@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#492514}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 765239ae75d08b5a09dddffa037747645e647abf