Call garbage collector when running with LSAN
LSAN is reporting objects that are allocated with PartitionAlloc
(replaced by Malloc when using LSAN). These object are kept alive
because some Blink (oilpan) objects are retaining them.

In this case, StyleRule is a oilpan garbage collected object but has
a member of type |CSSSelectorList| which is not garbage collected.

  see: https://cs.chromium.org/chromium/src/third_party/WebKit/Source/core/css/StyleRule.h?q=StyleRule&sq=package:chromium&dr=CSs&l=146


This CSSSelector is retaining the object reported as a leak:
  see: https://cs.chromium.org/chromium/src/third_party/WebKit/Source/core/css/CSSSelectorList.h?sq=package:chromium&dr=CSs&l=130


By calling the GC, the StyleRule are freed and it is releasing the memory allocated by PartitionAlloc.

Bug: 743134
Change-Id: Iea84298e4b087c6007124b436243c08bb381a4a9
Reviewed-on: https://chromium-review.googlesource.com/602070
Reviewed-by: Keishi Hattori <keishi@chromium.org>
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#492701}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b6018356b662d5a88cd30f906e4f041c68c8c3f6