Copy |server_address| in SyncResourceHandler
In ResourceDispatcher the |socket_address| of the SyncLoadResult is
copied into the SyncLoadResponse however in SyncResourceHandler this
field is not populated by copying the value from the ResourceResponse.
This patch fixes that and updates the single layout test that seems to
be affected.

Change-Id: I7e9962218da7c4fe188a340ec995118657eb62ef
Reviewed-on: https://chromium-review.googlesource.com/606793
Reviewed-by: Matt Menke <mmenke@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#492755}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 10820afc6a5b6462e5382ee518c74dcf763669be