Defer reporting FMP timestamp until SwapPromises complete
There are three corner cases with capturing the FMP swap timestamp that
this CL addresses:

1) If the provisional FMP timestamp is set, resulting in a queued
SwapPromise, and the 2-quiet timer fires before the SwapPromise
is fulfilled, then we do not report an FMP timestamp.

2) If the provisional FMP timestamp is set again while still waiting
for the previous SwapPromise, resulting in two outstanding
SwapPromises, and the 2-quiet timer fires in between SwapPromises,
then the wrong FMP swap timestamp will be reported. For example, the
following sequence of events can occur:
  a) Set provisional FMP & queue SwapPromise
  b) Set provisional FMP & queue SwapPromise
  c) Recieve SwapPromise result from (a)
  d) 2-quiet timer fires
  e) Recieve SwapPromise result from (b)
In this case, FMP is reported in (d) using the provisional FMP
timestamp from (b), but the SwapPromise timestamp corresponding to (a).

3) If the provisional FMP timestamp is less than the FCP timestamp when
the 2-quiet timer fires, we use FCP. However, if this timer fires
before the FCP SwapPromise has been fulfilled, then we report no FMP
swap timestamp.

To address these corner cases, reporting the FMP timestamp is deferred
if we either have any pending SwapPromises when the 2-quiet timer fires
or if we use the FCP timestamp but the corresponding FCP-swap timestamp
isn't set. In the first case, when the the number of outstanding
SwapPromises reaches 0, we report FMP. In the latter case, PaintTiming
informs the FirstMeaningfulPaintDetector when the swap timestamp is
available, and FMP is reported at that point.

Bug: 741961
Cq-Include-Trybots: master.tryserver.chromium.linux:linux_layout_tests_slimming_paint_v2
Change-Id: I202e668ab3295f7a1ee2bb10f925e54966b92c18
Reviewed-on: https://chromium-review.googlesource.com/599117
Reviewed-by: Shubhie Panicker <panicker@chromium.org>
Reviewed-by: Kunihiko Sakamoto <ksakamoto@chromium.org>
Reviewed-by: Steve Kobes <skobes@chromium.org>
Commit-Queue: Scott Haseley <shaseley@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#492835}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5f78548f700039bbfa3ee08a7a310a797f9e33f3