Add CHECK()s in ServiceWorkerGlobalScopeClient::From to investigate GC issue
According to the crash dump, ServiceWorkerGlobalScopeClient::From() looks like
returning null. But this should not return null. If it is actually returning
null, that means there is a GC issue in the CrossThreadPersistent mechanism or
in the Supplement mechanism. To investigate it, this CL add CHECK()s in
ServiceWorkerGlobalScopeClient::From().

Bug: 749930,728043
Change-Id: I813d44b84e2789b4f67dade847db9a6dc4c855bf
Reviewed-on: https://chromium-review.googlesource.com/606816
Reviewed-by: Keishi Hattori <keishi@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#492903}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dcc4a1093556ae395daef74b98620796be01e498