XSSAuditor: truncate form action attribute like other src-like attributes
Rescued from https://codereview.chromium.org/2868973003/ , which
somehow didn't land and escaped my notice when we switched to gerrit.

Adds a missing flag.

As a result of adding the flag, the issue that inspired the test
form-action-token-fragment.html becomes moot.  We're no longer
considering any part of the path/query/fragment, so they won't
influence behaviour.

See https://codereview.chromium.org/1179633002 for context. The
problem was the the auditor would sometimes fire on fragment
prefix matches (when right) and sometimes not (when wrong),
and might leak info as a result.

We fixed it last time by not firing until the entire fragment was right.
We fix it this time by always firing, even when the prefix is wrong.
As a result, the removed test as written would fail even
though things are safe.

Bug: 
Change-Id: I4cc05c882ff8b4b28b7d3d913916053124454b37
Reviewed-on: https://chromium-review.googlesource.com/600897
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#493117}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 80d2dd066eff7bac3dc0060bfe89067db48d17cf