Use CreateOrNull in CreateFloat32ArrayOrNull
When zero-initializing a Float32Array, use CreateOrNull instead of
Create.  We don't want to crash if we can't allocate the array.  This
matches the behavior when we're not initializing the array.  This also
preserves the contract that we should return null if we fail to create
the array.

Bug: 753291
Test: AudioBuffer/huge-buffer.html
Change-Id: I0635030e0c732561e979dbec7e1c123a93c73ec8
Reviewed-on: https://chromium-review.googlesource.com/608823
Commit-Queue: Raymond Toy <rtoy@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#493172}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5480d9713d544220c84c177f52ad1852ce92abf9