Use a separate enum for IntegrityMetadata::algorithm_.
This is a preparatory change for 607930. This separates the notion
of the integrity algorithm from a crypto hash function. The intent
is to introduce a signature-based scheme, and thus an integrity
method not based on a hash function.

Bug: 753349
Change-Id: Ibf1132249a66c4150bd66bbcfea6998607a35c62
Reviewed-on: https://chromium-review.googlesource.com/608974
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Daniel Vogelheim <vogelheim@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#493382}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ea0a536965d1709eecb29e9d13c09f013ff3adbe