Add edit commands to plugin context menu.
This CL adds support for executing "Cut", "Paste", and
"PasteAndMatchStyle" commands via the mouse and context menu in pepper
plugins. The plugin's context menu is built differently when focus is
within an editable text area: editable items are added as options and
regular page items are not shown.

Currently, the "SelectAll" command is disabled because it has not yet
been implemented, and spell check and writing direction sub menus are
not shown because support has not been added for them for plugins.

BUG=59266

Change-Id: Ib181f6dea29be3f524e5e19252ff12e34488ee16
Reviewed-on: https://chromium-review.googlesource.com/604827
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#493513}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4eb98ee11cded0a242fb3f072eb5715f4ce3e434