Fix flake in webusb/idlnarness.https.html
This patch fixes two related issues that caused this test to flake as a
result of a change to MojoInterfaceInterceptor. First, the idlharness.js
test does not explicitly wait for the Chromium-specific polyfill
implementation of navigator.usb.test to be loaded before trying to use
it. This worked anyways because there were enough asynchronous tasks
executed before it was needed in order to give it time to load.

The issue that itself caused the flake was that the Promise returned by
navigator.usb.test.initialize() resolves before the pipes to connect the
renderer to the mock service are established. This introduces the
a race between this initialization and the first fake device being
added that may cause its 'connect' event to never be fired. An explicit
call to navigator.usb.getDevices() is added to ensure that the pipes are
set up before we consider the test API initialized.

Bug: 751955
Change-Id: Ia874fe179b28bbeda0c987445be66dfd127e01cf
Reviewed-on: https://chromium-review.googlesource.com/610492
Reviewed-by: Vincent Scheib <scheib@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#493639}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9dabe0bb73adc9d777a7e4449b10e80c5d4a879c