Tcmalloc: Don't use MADV_FREE on linux
MADV_FREE is only available on Linux kernels >= 4.5.  Doing a
compile-time check to see if MADV_FREE is defined is insufficient.
This CL is an upstream of tcmalloc patch [1] which unconditionally
disables usage of MADV_FREE on Linux.  In the future, it would be
better to perform a runtime check so we can still use the feature when
it is available.

[1] https://github.com/gperftools/gperftools/commit/6f7a14f45e2e2e16fbbd250543f5758078e2f5f0

BUG=751812
R=wfh@chromium.org

Change-Id: Id1eceecd78ce26f74e8f061906902f7f3c2d650c
Reviewed-on: https://chromium-review.googlesource.com/614104
Reviewed-by: Will Harris <wfh@chromium.org>
Commit-Queue: Thomas Anderson <thomasanderson@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#494155}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4c2ef4092040a7a1fce76b6c543d26de4f2f0f29