Reland: Fix overlay scrollbar at huge scroll offsets
Raland note: This was landed previously in r491766 but reverted in
r491798 because the newly added test was failing on Android. On Android
the scrollbar margin was causing unexpected values in the test so this
patch modifies the test to a common value for margin (0).

The overlay scrollbar ThumbPosition method was losing precision at very
high scroll offsets. This would cause it to drift away from where we
actually show the thumb from the compositor so the hit testing would be
unreliable.

Looking at this method, there's no reason it should need to be any
different from the base class ScrollbarTheme::ThumbPosition which is
robust under high scroll offsets so the solution here is to simply
remove the override. Less code, less bugs.

Bug: 739668
Change-Id: Idaa61cd225b6cf67677dee364eb4fefed2ef4890
TBR: dtapuska@chromium.org, jbroman@chromium.org
Reviewed-on: https://chromium-review.googlesource.com/614614
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#494260}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d88adacb7803429007f81f62824b60c0f4b4358a