Set preload='' to more explicitly follow spec.
Currently, preload='' will default to 'auto' since it is marked as an
unsupported value. This means that when the default is set to
anything else (i.e. metadata for video), this will no longer adhere
to the spec. The spec[1] says, "The empty string is also a valid
keyword, and maps to the Automatic state."

[1] https://html.spec.whatwg.org/#attr-media-preload

Bug: 
Change-Id: Ia92031713d04bb2f3051fb4bc233d22153b775b9
Reviewed-on: https://chromium-review.googlesource.com/611104
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Commit-Queue: apacible <apacible@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#494266}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d796381b14b106a21aba7f18cb96aa9f79fb8c34