Add devicePixelRatio to PaintWorkletGlobalScope
In this CL, we add the devicePixelRatio property to the
PaintWorkletGlobalScope.idl. This property could be useful for 
developers to draw perfect graphics on high DPI devices.

A test is added to ensure that this property has the same value as the
window.devicePixelRatio.

Bug: 728599
Change-Id: Ia6179164989d4f564df06a87f1399adff19a0599
Reviewed-on: https://chromium-review.googlesource.com/602367
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Xida Chen <xidachen@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#494379}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 429cd7c7bf7d4f1d94bc570312ce95a8e86268e5