Add noise to the network quality estimates.
Up to 10% noise is added to the network quality estimates. The
noise is a function of the hostname.

Update the bucket to 50 msec (from 25 msec), and 50 kbps (from 25 kbps).

Put an upper limit on the RTT and downlink values reported.

Bug: 756029
Change-Id: Ida1afc14c5e855d8a06857c76d83b9912f663822
Reviewed-on: https://chromium-review.googlesource.com/604887
Commit-Queue: Tarun Bansal <tbansal@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Josh Karlin <jkarlin@chromium.org>
Reviewed-by: Martin Šrámek <msramek@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#494868}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f609e92a2ac7f44160901945b55337fcbf5e17a5