Fuchsia: remove calls to unported/unsupp. methods from FontCacheSkia
UI components, including font rendering, will not be ported initially
to Fuchsia. This CL ifdefs-out the portions of code which fail to
compile or link for Fuchsia headless shell builds, but the codepaths
are not meant to be executed at this point.

Bug: 707031
Change-Id: Icdedca4e109258e4416d65e07e36d81c45456444
Reviewed-on: https://chromium-review.googlesource.com/615844
Commit-Queue: Kevin Marshall <kmarshall@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#494911}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cd387344101f8f65cff0378d668d686c8080de10