[CSSParser] Allow parsers with observers to use streams.
Previously, we have CSSParserObservers which receive information about
the parsing process (e.g. where a particular rule starts or ends). The
way it worked for stylesheet parsing was that we first tokenized the
entirety of the input, storing token information in a vector. Then,
as we parsed, we queried the tokens for their location in the input
string (their offset) using the vector.

To get the same behaviour with streams, it's more difficult. Before,
offset information were available prior to parsing. Now, offset
information is only available as we parse. This requires some
refactoring to work.

In the meantime, we would like parsers with observers to still use
streams. Hence, this patch implements a workaround: when we have
observers, we tokenize everything before parsing as we did before.
We then create a stream over that tokenizer (technically no longer
a stream since everything's already tokenized) and use the same
parsing code path. Any time we call into the observer with a range,
we just convert the stream back into a range.

This means that until we implement proper streaming with observers,
parsing with observers will tokenize up front, whereas parsing
without observers will tokenize on demand (mostly).

For this to work, we also have to relax the requirement that
streams must use 'fresh' tokenizers with no consumed tokens,
since streams with observers requires using tokenizers that have
already consumed the entirety of the input. Turns out this doesn't
require any changes other than removing a DCHECK.

For reviewers:
- We add back the observer code that was in ConsumeDeclarationList
  but was deleted because observers didn't accept streams before.

Bug: 661854
Change-Id: Ifa53b2e7f2622edb371c74260fd24287d0fd4f89
Reviewed-on: https://chromium-review.googlesource.com/599467
Commit-Queue: Darren Shen <shend@chromium.org>
Reviewed-by: meade_UTC10 <meade@chromium.org>
Reviewed-by: Bugs Nash <bugsnash@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#494989}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7220fb9e1b829a9275b878b6fa499e55c4999cb9