Add DocumentMarkerController::MarkersIntersectingRange()
Currently, the only way to find the full list of markers intersecting a range is
to iterate through the nodes the range contains, and get the complete list of
nodes for each range from DocumentMarkerController, and do a linear search
through the list. This is inefficient if we have a text node with a lot of
markers, since we could be doing a binary search instead of a linear scan.

This CL introduces a DocumentMarkerController::MarkersIntersectingRange() method
which allows us to perform this operation much more efficiently.

Bug: 736181
Change-Id: Ib9c8036560a1ada642fc265a41e1bcadd5d5c949
Reviewed-on: https://chromium-review.googlesource.com/590948
Commit-Queue: Ryan Landay <rlanday@chromium.org>
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#495081}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 34710333ad9383a45f2fcf93d15b45add857c4bf