Document effective use of base::debug::Alias
Many uses of Alias(&var) end up being ineffective because var is a
pointer and the crash dump mostly only records the contents of the
thread stacks - therefore we end up with a pointer but not the memory
that it points to. This change documents this issue and gives examples
of effective use of base::debug::Alias.

BUG=756544,756589

Change-Id: I6095d5018e72c159bbe907c6294d7a48cf28fd82
Reviewed-on: https://chromium-review.googlesource.com/619112
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#495449}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 263423ff606d241556e0cadd05c2bfd21f0fbebb