Fix crash in CountDeprecationWarning
The code to count cross-origin uses of features which are going to be
under the control of Feature Policy wasn't checking whether Feature
Policy itself was enabled before asking the frame to check its policy.
This caused a DCHECK to fire in Frame.cpp in that case.

Without Feature Policy, we can't meaningfully check whether the usage
of the feature would have been declined in every case, so this change
causes CountDeprecationWarning to return early. This may introduce some
undercounting, but since Feature Policy is now enabled by default, that
should be minimal (and this undercounting is not new; it was just
previously represented by crashes instead).

Bug: 752762
Change-Id: Ia36a2f7c85de8cedc4a386803d428e9d351db012
Reviewed-on: https://chromium-review.googlesource.com/619366
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Ian Clelland <iclelland@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#495543}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fb86b10bda0326d2e05c3f5f4060120613f91a76