[CSSParser] Fix CSSParserObserverWrapper workaround.
In [1], we introduced a temporary workaround that allowed
CSSParserObserverWrapper to take streams as well as ranges. However,
there was a bug in that patch. We used MakeRangeToEOF() to convert
the stream to a range, but that means the converted range would end
at EOF. While this was fine for CSSParserObserverWrapper::StartOffset,
which only used the the starting point of the range, it was broken for
CSSParserObserverWrapper::EndOffset, which used the ending point.

It's quite tricky to get the "end point" of a stream, since it's not
well defined. We looked at usages of CSSParserObserverWrapper::EndOffset
and saw that they were only called when parsing a range is successful
(i.e. the range was fully consumed). In that scenario, the end point of
the range is equal to the start point of the range. Hence, the correct
end point is really just the position of the stream when we call
EndOffset.

Since the observer wrapper only looks at either the starting or ending
point of the range, we can just pass it an empty range with starting
and ending points set to the correct value (the current position of
the stream).

[1] https://chromium-review.googlesource.com/c/599467

CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.win:win_optional_gpu_tests_rel

Bug: 661854
Change-Id: I191dbf6946dc7b3b1a6e81b11d2cf95ef3b72b33
Reviewed-on: https://chromium-review.googlesource.com/622331
Reviewed-by: nainar <nainar@chromium.org>
Commit-Queue: Darren Shen <shend@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#496089}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5819e64ac263e30c62fe1e2b3ebb565665ad3035