Remove unused/redundant method Port.skipped_layout_tests.
The list returned by Port.skipped_layout_tests was added to the list of
tests with skip expectations, but nothing besides the unit tests are
actually using this now.

This CL removes this method and updates the tests.

In general, the set of tests with skip expectations are those
with Skip or WontFix test expectation lines, or those explicitly
skipped with --ignore-tests.

Change-Id: If8dbd02366e8dde143edc604061341d6c2d2d276
Reviewed-on: https://chromium-review.googlesource.com/624657
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#496145}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 70f0ea0fadbef5428cf59124f41b35f579d84846