Include shadow host in elements invalidated by type ruleset invalidations.
When we do a ruleset invalidation for a shadow scope, we schedule an
invalidation set for type selectors on the shadow root node.
This set might also contain tag names for type selectors in :host().
In order to invalidate host elements properly, check if the host's tag name is
contained in any of the type invalidation sets and mark it for style recalc.

Bug: 754191
Change-Id: I92b5a5dfacb86e65ff52247db34fbc9fa0b2101d
Reviewed-on: https://chromium-review.googlesource.com/618069
Commit-Queue: Eriko Kurimoto <elkurin@google.com>
Reviewed-by: Rune Lillesveen <rune@opera.com>
Reviewed-by: Takayoshi Kochi <kochi@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#496561}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a021e41cfd7d1a56bbc30e1d5346fc44d728d7e6