Make CharacterIndexAtPointWithPinchZoom load Ahem
This test uses the Ahem font but on machines/bots without this font
it will fallback. The expectations were set for the fallback without
realizing it. This causes the test to fail on a system with Ahem
installed.

This patch calls LoadAhem in the test and resets the expectations.

Bug: 749836
Change-Id: I7c0854f8d45bbadf1b88c27ab4964b2dac312058
Reviewed-on: https://chromium-review.googlesource.com/627378
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#496585}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 57f5b834928ff596e8c5520d532ac8f22706886f