Remove unnecessary condition in LayoutBlock
Basically, positioned object can't be the root of a layout tree.
So we do not need to check nullptr for "o" in while loop
because of parent always exists.

This patch is just for cleaning up the codes.

BUG=752371

Change-Id: I4b1f10c9b723b0736fa90fd488ca03973c48667e
Reviewed-on: https://chromium-review.googlesource.com/601628
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Stephen Chenney <schenney@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#496706}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bf20977ca2b1af2eff9769679dc45015711c0b57