Create scrollable cc:Layers from ScrollHitTest display items
This patch changes PaintArtifactCompositor to create layers specifically
for scrolling that are sized to the scroll container's bounds. These
layers are inserted in the correct paint order because they are
created from ScrollHitTestDisplayItems.

This patch has a change in the property nodes used for the scroll hit
test layer so that the layer is no longer under the scroll or overflow
clip nodes. This ensures crbug.com/753124 will be fixable in the future
by not clipping the hit test region and including borders. To do this,
the scroll node is now passed on the scroll hit test display item.

Bug: 738613
Cq-Include-Trybots: master.tryserver.chromium.linux:linux_layout_tests_slimming_paint_v2
Change-Id: If0ade80cfd233384f4b74923148b2d79eb4b41d6
Reviewed-on: https://chromium-review.googlesource.com/609317
Reviewed-by: Tien-Ren Chen <trchen@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: Philip Rogers <pdr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#496907}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: eef5607b471f955590278d5e44762c5206272caa