Fix copy paste error
In field.tmpl, template for inherit property comparison has a
mistake where an unknown field field_to_compare get called. 
Fixing it to subgroup as it was intended. This will cause 
a compile error if an inherited property is in a nested 
group, which means 
{
name: "foo",
inherited: true,
field_group: "a->b->c"
}
There is no changes to the current ComputedStyleBased 
because we have not make inherited properties to have 
nested groups. 

This error was found when we give nested group to inherited
properties.

Bug: 
Change-Id: Icf2a4681b91c2a9d03675bf2a0e54526c454c2bf
Reviewed-on: https://chromium-review.googlesource.com/631297
Commit-Queue: Minh-Duc Nguyen <nmduc@google.com>
Reviewed-by: nainar <nainar@chromium.org>
Reviewed-by: Darren Shen <shend@chromium.org>
Reviewed-by: meade_UTC10 <meade@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#497293}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fe9f6912db7ba5e7efde315d7152a96f0019f2e1