Rename DocumentMarkerListEditor to SortedDocumentMarkerListEditor
I'm going to introduce a new class, UnsortedDocumentMarkerListEditor, for marker
types that we store unsorted because they can overlap. Currently, suggestion
markers are the only markers we have of this type, so the logic is implemented
directly in SuggestionMarkerListImpl, but I also want to make
CompositionMarkerListImpl use the same logic to fix crbug.com/752648. Therefore,
we need to rename DocumentMarkerListEditor to clarify that it should only be
used for MarkerListImpls that store markers in sorted order.

Bug: 752648
Change-Id: I796e8d23cb75a3c45e6c1639ad67bc266668910f
Reviewed-on: https://chromium-review.googlesource.com/636181
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Commit-Queue: Ryan Landay <rlanday@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#497854}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1b051d44ad4685de6d32a65f554d0a42bc0e6170