Introduce core/dom/events directory, and move related files to there
This is one of the efforts to organize core/dom files.  See
https://docs.google.com/spreadsheets/d/1OydPU6r8CTj8HC4D9_gVkriJETu1Egcw2RlajYcw3FM/edit?usp=sharing
for details.

core/dom/events directory should be used as an implementation of a general
mechanism for event dispatching, which is defined in "DOM Events" section in
DOM Standard.

This CL hasn't moved implementations of each event type, which are defined in
other specifications, such as UI Events, from core/events directory.  We might
want to merge core/events and core/input in the future, but that should be done
in another CL.

Bug: 738794
Change-Id: Ibbdab85f4484ffd1a7029e355f880c5227c8ad24
Reviewed-on: https://chromium-review.googlesource.com/640275
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Hayato Ito <hayato@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#498066}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c4b283d57e4e87f69d510bbc1fe21c6e173fa1fa