Fix crash when indexing into column collapse vector
Take the minimum of the number of columns and the number of columns a
cell is spanning. Before, we did'nt take the minimum, so a cell set to
span more columns than the actual number of columns resulted in an
error.

Bug: 758503
Change-Id: I3257bf3313f3ad09850a531c9ca1c5511d94b0f1
Reviewed-on: https://chromium-review.googlesource.com/642119
Reviewed-by: Emil A Eklund <eae@chromium.org>
Reviewed-by: David Grogan <dgrogan@chromium.org>
Commit-Queue: David Grogan <dgrogan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#498270}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b25ea97e412367de1c639b092c40a5b927fcb8aa