Fix LazyLineBreakIterator not to crash for null strings
This patch fixes LazyLineBreakIterator not to crash for null strings
when it checks Is8Bit() in NextBreakablePosition().

This has never been a case but we started seeing cases where this is
the most likely cause on Android.

This is a speculative fix.

Bug: 756624
Change-Id: Ifd5a9d2c3c0034e916c2671ea1ed14ae975c9adc
Reviewed-on: https://chromium-review.googlesource.com/636789
Reviewed-by: Emil A Eklund <eae@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#498403}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7a53f297789344cbf5df695fa46c251ad96b356d