Add UnsortedDocumentMarkerListEditor::ShiftMarkersContentIndependent()
This method is not used by SuggestionMarkerListImpl, and therefore was not
included in the original CLs to pull out SuggestionMarkerListImpl's methods into
UnsortedDocumentMarkerListEditor. However, we do need it to implement
CompositionMarkerListImpl on top of UnsortedDocumentMarkerListEditor, so I'm
adding it here.

Bug: 752648
Change-Id: Ibd006cdfb94a0060a9fc2c0962b92be9426e6528
Reviewed-on: https://chromium-review.googlesource.com/644145
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Commit-Queue: Ryan Landay <rlanday@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#498638}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5916acd85b148e534152b329ecc682bdcb7cd427