Blob interface: remove close() / isClosed
Chrome had an experimental implementation of a close() method (and
corresponding isClosed attribute) to allow web apps to explicitly
drop the resources associated with a Blob when no longer needed.
This was in the spec for a while but due to flaws [1] and lack of
implementation by other browsers, it was removed from the spec [2].

Remove the method/attribute and associated tests. It required the
"experimental" flag to use, so there is no risk of web breakage.

[1] https://github.com/w3c/FileAPI/issues/10
[2] https://github.com/w3c/FileAPI/pull/68

Bug: 344820
Change-Id: I6c8ff1f73590f40b68f2b862e15a9db3936bc0f1
Reviewed-on: https://chromium-review.googlesource.com/638915
Reviewed-by: Philip Jägenstedt <foolip@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Commit-Queue: Joshua Bell <jsbell@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#499019}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3cd8b6d6b28ef6e195c2d592520d37337a19f09c