images: Use TimeDelta in the image stack from decoders to BitmapImage.
This patch ensures that we don't accidentally use wrong time units when
using deltas. Decoders report this in milliseconds, but we use this in
seconds in BitmapImage. This patch makes everything in the stack use
TimeDelta, and only convert it to seconds when needed.

R=chrishtr@chromium.org, khushalsagar@chromium.org

Change-Id: I60beddc3fc966c02a7812050d0b0712586745206
Reviewed-on: https://chromium-review.googlesource.com/644114
Commit-Queue: Vladimir Levin <vmpstr@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#499333}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bd22c1ecd5ca1bcc7fe7c8d14156cb6502983f6f