Re-land: Add more AOM Phase 2 event listeners.
Originally landed: r499011
Reverted: 499164

The issue was just a layout test timing out on Windows.
The fix is to split it into more files with fewer
individual tests per file.

This change adds support for these new events:
  accessiblecontextmenu
  accessibledecrement
  accessiblefocus
  accessibleincrement
  accessiblescrollintoview

The pattern for all of these is the same - when the
action is triggered on an AXObject, we first try to fire
an AOM event, and if that's not handled, we want to fall
back on the default behavior.

To make that more readable, this change renames all of
the action interfaces in AXObject to each consist of a
pair of methods - one public method to be triggered from
the outside, and an internal method to handle the
native fallback behavior.

Bug: 746524
Change-Id: Icbe20c83898dd588c23b81ddcf17276a203369ea
TBR: aboxhall@chromium.org
Reviewed-on: https://chromium-review.googlesource.com/648360
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#499452}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2b781f033b8d428dbe5d3d36455977fcdc9928da