ES6 Modules: Change ModuleScriptFetcher to an abstract class
For cleanup, this CL changes ModuleScriptFetcher to an abstract class, and
makes {Document,Worklet}ModuleScriptFetcher directly inherit it.

Before this CL, ModuleScriptFetcher is a concrete class used for module loading
on the Document context, and WorkletModuleScriptFetcher inherits
ModuleScriptFetcher for the Worklet context. This worked well until [1] because
WorkeltModuleScriptFetcher depended on functions of ModuleScriptFetcher (e.g.,
implementation of ResourceOwner). However, after the change,
WorkletModuleScriptFetcher no longer depends on them and doesn't have to
inherit it.

To simplify the code, this CL...
- changes ModuleScriptFetcher to an abstract class.
- introduces DocumentModuleScriptFetcher as implementation of
  ModuleScriptFetcher for the Document context. This is the same as the previous
  ModuleScriptFetcher.
- makes WorkletModuleScriptFetcher directly inherit ModuleScriptFetcher as
  opposed to DocumentModuleScriptFetcher.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/623044

Bug: 727194
Change-Id: I42631f5c21d8473b9a121883b4a5b027c56089cd
Reviewed-on: https://chromium-review.googlesource.com/641297
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#499612}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d22d2df8014490b385e2bc2700a11a0bf3201146