[css-grid] grid shorthand does not reset gutter properties
The CSS WG has resolved that the "grid" shorthand shouldn't reset
the gutter properties anymore:
https://github.com/w3c/csswg-drafts/issues/1036

This was originally implemented in r370102, so this is somehow
a revert of that change.
The patch adds a new test to WPT verifying the expected behavior.

BUG=761905
TEST=external/wpt/css/css-grid-1/grid-definition/grid-shorthand-001.html

Change-Id: Ib8b0591a0132c6520c3cfd8f13fb2cf2a33e1be0
Reviewed-on: https://chromium-review.googlesource.com/650406
Reviewed-by: Sergio Villar <svillar@igalia.com>
Reviewed-by: Philip Jägenstedt <foolip@chromium.org>
Reviewed-by: meade_UTC10 <meade@chromium.org>
Reviewed-by: Javier Fernandez <jfernandez@igalia.com>
Commit-Queue: Manuel Rego Casasnovas <rego@igalia.com>
Cr-Original-Commit-Position: refs/heads/master@{#499901}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 850c817c6a46e5b3a420c02898e2adc2376d91fc